refactor: Use ifdef instead of different source folders
