--- src/main/3.0_and_up/scala/dev/mauch/spark/excel/v2/ExcelDataSource.scala
@@ -1,49 +1,0 @@
-/*
- * Copyright 2022 Martin Mauch (@nightscape)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package dev.mauch.spark.excel.v2
-
-import org.apache.spark.sql.connector.catalog.Table
-import org.apache.spark.sql.execution.datasources.FileFormat
-import org.apache.spark.sql.execution.datasources.v2.FileDataSourceV2
-import org.apache.spark.sql.types.StructType
-import org.apache.spark.sql.util.CaseInsensitiveStringMap
-
-/** Derived from Spark own CSV implementation
-  */
-class ExcelDataSource extends FileDataSourceV2 {
-
-  override def fallbackFileFormat: Class[_ <: FileFormat] = classOf[ExcelFileFormat]
-
-  override def getTable(options: CaseInsensitiveStringMap): Table = {
-    val paths = getPaths(options)
-    val tableName = getTableName(options, paths)
-    val optionsWithoutPaths = getOptionsWithoutPaths(options)
-    ExcelTable(tableName, sparkSession, optionsWithoutPaths, paths, None)
-  }
-
-  override def getTable(options: CaseInsensitiveStringMap, schema: StructType): Table = {
-    val paths = getPaths(options)
-    val tableName = getTableName(options, paths)
-    val optionsWithoutPaths = getOptionsWithoutPaths(options)
-    ExcelTable(tableName, sparkSession, optionsWithoutPaths, paths, Some(schema))
-  }
-
-  /** The string that represents the format that this data source provider uses
-    */
-  override def shortName(): String = "excel"
-
-}
