--- src/main/3.3_and_up/scala/dev/mauch/spark/excel/v2/ExcelScanBuilder.scala
@@ -1,40 +1,0 @@
-/*
- * Copyright 2022 Martin Mauch (@nightscape)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package dev.mauch.spark.excel.v2
-
-import org.apache.spark.sql.SparkSession
-import org.apache.spark.sql.connector.read.Scan
-import org.apache.spark.sql.execution.datasources.PartitioningAwareFileIndex
-import org.apache.spark.sql.execution.datasources.v2.FileScanBuilder
-import org.apache.spark.sql.internal.connector.SupportsPushDownCatalystFilters
-import org.apache.spark.sql.types.StructType
-import org.apache.spark.sql.util.CaseInsensitiveStringMap
-
-case class ExcelScanBuilder(
-  sparkSession: SparkSession,
-  fileIndex: PartitioningAwareFileIndex,
-  schema: StructType,
-  dataSchema: StructType,
-  options: CaseInsensitiveStringMap
-) extends FileScanBuilder(sparkSession, fileIndex, dataSchema)
-    with SupportsPushDownCatalystFilters {
-
-  override def build(): Scan = {
-    ExcelScan(sparkSession, fileIndex, dataSchema, readDataSchema(), readPartitionSchema(), options, pushedDataFilters)
-  }
-
-}
